Move `StandardGate` definition into its own module
This is a staging commit, splitting up the giant monolithic
`qiskit_circuit::operations` to put `StandardGate` in its own module.
The code is unchanged, other than imports.

This is preparatory for moving more code out of `quantum_info` that
connects `StandardGate` to base `quantum-info` objects.

Unfortunately I couldn't make this register in `git` as a giant move, so
the best way to review this commit is to verify that the diff of the
moved block (lines 1163 to 3084) in `crates/circuit/src/operations.rs`
in the parent of this commit[^1] is minimal.  For example:

```bash
( git show 7ceda6324b3d:crates/circuit/src/operations.rs |
        head -n 3084 |
        tail -n+1163 ) |
    diff - crates/circuit/src/standard_gate/mod.rs
```

shows

```text
0a1,27
> // This code is part of Qiskit.
> //
> // (C) Copyright IBM 2026
> //
> // This code is licensed under the Apache License, Version 2.0. You may
> // obtain a copy of this license in the LICENSE.txt file in the root directory
> // of this source tree or at https://www.apache.org/licenses/LICENSE-2.0.
> //
> // Any modifications or derivative works of this code must retain this
> // copyright notice, and modified files need to carry a notice indicating
> // that they have been altered from the originals.
>
> use crate::circuit_data::{CircuitData, PyCircuitData};
> use crate::operations::{Operation, Param, add_param, clone_param, multiply_param, radd_param};
> use crate::{Qubit, gate_matrix, impl_intopyobject_for_copy_pyclass, imports};
>
> use ndarray::{Array2, aview2};
> use num_complex::Complex64;
> use smallvec::{SmallVec, smallvec};
> use std::f64::consts::PI;
>
> use numpy::{IntoPyArray, PyArray2};
> use pyo3::prelude::*;
> use pyo3::types::{IntoPyDict, PyList, PyTuple};
>
> const FLOAT_ZERO: Param = Param::Float(0.0);
>
1920,1922d1946
<
< const FLOAT_ZERO: Param = Param::Float(0.0);
<
```

[^1]: 7ceda6324b3d: Revert "Support commutation check between Pauli-based gates and standard gates (#15488)" (#15906)
