+++ crates/circuit/src/lib.rs
@@ -36,6 +36,7 @@ pub mod parameter;
 pub mod parameter_table;
 pub mod register_data;
 pub mod slice;
+pub mod standard_gate;
 pub mod var_stretch_container;
 mod variable_mapper;
 pub mod vf2;
